pyproject.toml
[project]
name = "fredo"
version = "0.1.0"
description = "A CLI tool for managing and running code snippets with GitHub Gist integration"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0,<0.13.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "prompt-toolkit>=3.0.0",
    "thefuzz[speedup]>=0.22.0",
    "pygithub>=2.3.0",
    "pygments>=2.17.0",
    "pydantic>=2.7.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "freezegun>=1.4.0",
]

[project.scripts]
fredo = "fredo.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
